Add JWT module with decoding and verification shards
